WIP API
* moved API into an api/ directory
* added empty save, get_my_groups, gey_my_projects
* working on delete method to make it clean and intuitive
* added empty PrimaryNode base class
