+++ src/nodes/primary_nodes/PrimaryNode.py
@@ -1,0 +1,2 @@
+class PrimaryNode:
+    pass
